dspy_boss/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dspy-boss"
version = "0.1.0"
description = "Generalized DSPy boss with state machine, MCP loading, agents, and async task manager"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7,<3",
  "pyyaml>=6.0",
  "typer>=0.12",
  "rich>=13.7",
  "anyio>=4.4",
  "httpx>=0.27",
]

[project.scripts]
dspy-boss = "dspy_boss.cli:app"
